Fixed compile error on gcc. Reviewed jl.

+++ objects/util/base/include/time_vector.h
@@ -78,7 +78,7 @@ namespace objects {
 
 
             //! Current index into the array.
-            unsigned int mPos;
+            size_t mPos;
 
             //! Parent container.
             const TimeVectorBase* mParent;
